Add locking
This adds locks to Goombler.java that prevent the per-goombler race condition.

This still leaves us open to a race condition in the "global" GoombleAccount balance because multiple threads acting on different goomblers can run into each other in the global balance. I added an `AtomicInteger` to `GoombleAccount` to highlight the differences.

I also removed all the references to the ultimately unused region numbers.