Make a Java version that always locks
I was a little concerned that constantly checking whether to use locks or not was unfairly slowing down the Java version, so I separated that out into two versions: one that always locks, and one that never locks. This didn't substantially change any of the timing, presumably because the blocking is what's killing us.