Experiment with crossbeam
This still doesn't work, but it at least compiles. It turns forever, though, because the scope that reads the channels is waiting for all the channels to be empty before anything is ever written to them.

Not entirely sure how I fix this, but I feel like I'm learning useful things.