Add sleeps to the Clojure version & Rust version with locks.
I realized in getting the Rust channels to work, that I had *no* sleeps on either the Rust code or the Clojure code, which was perhaps why things were so much faster there. And…yup – that was indeed the case.

Adding this sleep to the Clojure code slows it down a *ton*, making it the slowest version by a wide margin. I think the problem is that it's in a `dosync` which will keep retrying until it gets through the whole thing without someone else cutting in before it finishes. The odds of that go way down with the sleep here, so I think we retry a _lot_.

Adding sleep to the Rust code makes it essentially comparable speed with the Java code. Oops.