Get the Rust version with channels working
This gets the Rust version with channels working!

I must confess I still only half understand ownership and mutability and references in Rust, and there was a lot of flailing and add/removing `&` and `mut` here and there to get this to work, but it does work, and it's actually 25-30% faster than any of the other versions. I suspect this is because there's essentially no locking here. In general things only block if they are attempting to read from an empty channel, and that probably doesn't happen often until the end when all the channels close up. Threads will get pulled off the CPU here and there to make room for other threads, but probably less than with all the lock checking in the other approaches.