Replace use of random sleep with random fib calls
This replaces the random sleep calls in all four versions with random calls to fib so that the threads are actually doing some work.

I'm not sure if I want this change or not. It moves the timing away from just the locking to doing actual computation, so I think it's a distraction, TBH.